requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
itsdangerous==2.2.0
python-dotenv==1.0.1
firebase-admin==6.7.0
google-cloud-firestore==2.21.0